requirements.txt
flask
keras
numpy
Pillow
tensorflow